Update templates to pass seo metadata
All pages should now pass appropriate values for title, description, and
image metadata to the seo component. The least invasive change was to
pass the props on through the layout component but the more correct
might be to directly reference the seo component in the templates.

Setting the images to 1200 for now since that's the recommended width
for social cards but it may need some tweaking.

This fixes #124