Specify the base site url for the different environments
The facebook and twitter crawlers expect the metadata links to use
absolute urls so in order to be able test to the seo changes we needed
to configure the base url and the quickest way seemed to hardcode it for
staging and this particular pull request in the `netlify.toml`.
These could be updated as needed for testing.

Ideally we'd be able to leverage the netlify DEPLOY_URL environment
variable but I wasn't able to get it working.