Improve SEO for individual pages (#127)
* Update seo component to better handle individual pages

Now takes into account specific pages and templates can pass in custom
metadata. There are still properties that need to be set properly but
this addresses the most immediate needs.

* Update templates to pass seo metadata

All pages should now pass appropriate values for title, description, and
image metadata to the seo component. The least invasive change was to
pass the props on through the layout component but the more correct
might be to directly reference the seo component in the templates.

Setting the images to 1200 for now since that's the recommended width
for social cards but it may need some tweaking.

This fixes #124

* Specify the base site url for the different environments

The facebook and twitter crawlers expect the metadata links to use
absolute urls so in order to be able test to the seo changes we needed
to configure the base url and the quickest way seemed to hardcode it for
staging and this particular pull request in the `netlify.toml`.
These could be updated as needed for testing.

Ideally we'd be able to leverage the netlify DEPLOY_URL environment
variable but I wasn't able to get it working.

* add featuredimage to cms admin fields. Make a featureImage1200 fragment for more reausability.

* add contain to handle images that are _not_ the correct size.

* fix broken quotes

* use constant instead of string for fit param.

Co-authored-by: Garrett Van Gotten <garrett.van-gotten@c2fo.com>