Add Visual Studio Code development container
Using the "Node.js 12 & JavaScript" template provided by VS Code, create
a development container configuration that sets up a Dockerized
development environment for working on the site.

To ensure the correct Node.js version is installed, Docker needs to copy
over `.nvmrc`. To limit the build context (files) sent to Docker while
building the container, a `.dockerignore` file was added in the root of
the project.